Inline API key list actions
